feat: improved UI of order creation tab
