+++ src/ui/helpers.rs
@@ -1,10 +1,10 @@
 use crate::models::AdminDispute;
 use chrono::DateTime;
 use mostro_core::prelude::UserInfo;
-use ratatui::layout::{Constraint, Flex, Layout, Rect};
+use ratatui::layout::{Constraint, Direction, Flex, Layout, Rect};
 use ratatui::style::{Color, Modifier, Style};
 use ratatui::text::{Line, Span};
-use ratatui::widgets::{ListItem, Paragraph};
+use ratatui::widgets::{Borders, ListItem, Paragraph};
 use std::fs::{self, OpenOptions};
 use std::io::Write;
 
@@ -96,6 +96,108 @@ pub fn render_help_text(f: &mut ratatui::Frame, area: Rect, prefix: &str, key: &
         ]))
         .alignment(ratatui::layout::Alignment::Center),
         area,
+    );
+}
+
+/// Render a pair of centered YES/NO buttons inside the given area.
+/// `selected_button = true` highlights YES, `false` highlights NO.
+pub fn render_yes_no_buttons(
+    f: &mut ratatui::Frame,
+    area: Rect,
+    selected_button: bool,
+    yes_label: &str,
+    no_label: &str,
+) {
+    let button_width = 15;
+    let separator_width = 1;
+    let total_button_width = (button_width * 2) + separator_width;
+
+    let button_x = area.x + (area.width.saturating_sub(total_button_width)) / 2;
+    let centered_button_area = Rect {
+        x: button_x,
+        y: area.y,
+        width: total_button_width.min(area.width),
+        height: area.height,
+    };
+
+    let button_chunks = Layout::new(
+        Direction::Horizontal,
+        [
+            Constraint::Length(button_width),
+            Constraint::Length(separator_width),
+            Constraint::Length(button_width),
+        ],
+    )
+    .split(centered_button_area);
+
+    // YES button
+    let yes_style = if selected_button {
+        Style::default()
+            .bg(Color::Green)
+            .fg(Color::Black)
+            .add_modifier(Modifier::BOLD)
+    } else {
+        Style::default()
+            .fg(Color::Green)
+            .add_modifier(Modifier::BOLD)
+    };
+
+    let yes_block = ratatui::widgets::Block::default()
+        .borders(Borders::ALL)
+        .style(yes_style);
+    f.render_widget(yes_block, button_chunks[0]);
+
+    let yes_inner = Layout::new(Direction::Vertical, [Constraint::Min(0)])
+        .margin(1)
+        .split(button_chunks[0]);
+
+    f.render_widget(
+        Paragraph::new(Line::from(vec![Span::styled(
+            yes_label,
+            Style::default()
+                .fg(if selected_button {
+                    Color::Black
+                } else {
+                    Color::Green
+                })
+                .add_modifier(Modifier::BOLD),
+        )]))
+        .alignment(ratatui::layout::Alignment::Center),
+        yes_inner[0],
+    );
+
+    // NO button
+    let no_style = if !selected_button {
+        Style::default()
+            .bg(Color::Red)
+            .fg(Color::Black)
+            .add_modifier(Modifier::BOLD)
+    } else {
+        Style::default().fg(Color::Red).add_modifier(Modifier::BOLD)
+    };
+
+    let no_block = ratatui::widgets::Block::default()
+        .borders(Borders::ALL)
+        .style(no_style);
+    f.render_widget(no_block, button_chunks[2]);
+
+    let no_inner = Layout::new(Direction::Vertical, [Constraint::Min(0)])
+        .margin(1)
+        .split(button_chunks[2]);
+
+    f.render_widget(
+        Paragraph::new(Line::from(vec![Span::styled(
+            no_label,
+            Style::default()
+                .fg(if !selected_button {
+                    Color::Black
+                } else {
+                    Color::Red
+                })
+                .add_modifier(Modifier::BOLD),
+        )]))
+        .alignment(ratatui::layout::Alignment::Center),
+        no_inner[0],
     );
 }
