+++ src/ui/key_handler/esc_handlers.rs
@@ -13,7 +13,7 @@ pub fn handle_esc_key(app: &mut AppState) -> bool {
             app.mode = default_mode.clone();
             true
         }
-        UiMode::UserMode(UserMode::ConfirmingOrder(form)) => {
+        UiMode::UserMode(UserMode::ConfirmingOrder { form, .. }) => {
             // Cancel confirmation, go back to form
             app.mode = UiMode::UserMode(UserMode::CreatingOrder(form.clone()));
             true
